+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Fix nils displayed in keypad popup
 * Fix C-S- and C-M-S- in keypad
 * Eval `meow-motion-overwrite-define-key` multiple times cause invalid remap
+* Set `undo-tree-enable-undo-in-region` for undo-tree automatically
 
 ## 1.0.1 (2021-11-30)
 ### Bugs fixed